Fix non-passive event listener warnings
Running this project's code in Chrome results in two console warnings at
the "Verbose" level:
```
Added non-passive event listener to a scroll-blocking 'touchmove' event.
Added non-passive event listener to a scroll-blocking 'touchstart' event.
```

The rest of the warning message is:
> Consider marking event handler as 'passive' to make the page more
> responsive. See https://www.chromestatus.com/feature/5745543795965952

It is necessary to use feature detection
(`supportsPassiveEventListeners`) here because older browsers do not
support an object as a third parameter of `addEventListener`.

Signed-off-by: Adam Vigneaux <avigneaux@6river.com>